Add files via upload
to use it, cd to the file where Project Pro is, in terminal, run "npm run dev", after connecting to mongodb, you can go to "http://localhost:3000" and it should be working. For this version, you should be able to sign in with google account, access home page and search nearby page, and all the search function inside that two pages should work. "Latest News" page is still under development, I should be able to get it done by tomorrow.